Update console app
