+++ MelonChart.sln
@@ -7,13 +7,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{3D03FD67-99F
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MelonChart", "src\MelonChart\MelonChart.csproj", "{8C61EBBC-B2B6-4C3D-A7AA-2D9569EAE8DC}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{ACA1464B-2D70-4B57-B6A6-2E399931AC0A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MelonChart.ConsoleApp", "sample\MelonChart.ConsoleApp\MelonChart.ConsoleApp.csproj", "{FA517387-3495-4649-81D3-AF51A9A683B8}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{8CF3DBA1-84DB-4274-86EA-3BB54B7262C0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MelonChart.Tests", "test\MelonChart.Tests\MelonChart.Tests.csproj", "{48FBE532-F50C-4CE5-A61D-86B5921B9ACE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MelonChart.Tests", "test\MelonChart.Tests\MelonChart.Tests.csproj", "{48FBE532-F50C-4CE5-A61D-86B5921B9ACE}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{ACA1464B-2D70-4B57-B6A6-2E399931AC0A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MelonChart.ConsoleApp", "samples\MelonChart.ConsoleApp\MelonChart.ConsoleApp.csproj", "{FA517387-3495-4649-81D3-AF51A9A683B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,22 +25,22 @@ Global
 		{8C61EBBC-B2B6-4C3D-A7AA-2D9569EAE8DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8C61EBBC-B2B6-4C3D-A7AA-2D9569EAE8DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8C61EBBC-B2B6-4C3D-A7AA-2D9569EAE8DC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FA517387-3495-4649-81D3-AF51A9A683B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FA517387-3495-4649-81D3-AF51A9A683B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FA517387-3495-4649-81D3-AF51A9A683B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FA517387-3495-4649-81D3-AF51A9A683B8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{48FBE532-F50C-4CE5-A61D-86B5921B9ACE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{48FBE532-F50C-4CE5-A61D-86B5921B9ACE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{48FBE532-F50C-4CE5-A61D-86B5921B9ACE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{48FBE532-F50C-4CE5-A61D-86B5921B9ACE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FA517387-3495-4649-81D3-AF51A9A683B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FA517387-3495-4649-81D3-AF51A9A683B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FA517387-3495-4649-81D3-AF51A9A683B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FA517387-3495-4649-81D3-AF51A9A683B8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{8C61EBBC-B2B6-4C3D-A7AA-2D9569EAE8DC} = {3D03FD67-99F7-4245-83C4-230456C5D9F5}
-		{FA517387-3495-4649-81D3-AF51A9A683B8} = {ACA1464B-2D70-4B57-B6A6-2E399931AC0A}
 		{48FBE532-F50C-4CE5-A61D-86B5921B9ACE} = {8CF3DBA1-84DB-4274-86EA-3BB54B7262C0}
+		{FA517387-3495-4649-81D3-AF51A9A683B8} = {ACA1464B-2D70-4B57-B6A6-2E399931AC0A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {34032734-030B-4982-8F04-9138EFCDF681}
