+++ samples/MelonChart.ConsoleApp/Program.cs
