+++ samples/MelonChart.ConsoleApp/Properties/launchSettings.json
