+++ samples/MelonChart.ConsoleApp/Services/IMelonChartService.cs
